Clean up: move integration tests into the test module.
Wary of file proliferation, the integration tests from module
test_file_extraction_integration_tests were placed into a
separate class in module test_file_extraction and module
test_file_extraction_integration_tests was removed.
